Fix 6 bugs: UTF-8 panics, error swallowing, f64→f32 threshold, cosine epsilon
- Fix UTF-8 string slicing panics in compiler.rs (2 sites) and
  graph_batch.rs — use char_indices() to find safe truncation points
  instead of naive byte offsets that panic on multi-byte content
- Propagate errors in ontology.rs (serde_json) and graph_batch.rs
  (chunk fetch) instead of silently defaulting to empty values
- Fix f64→f32 precision loss in pg_resolver.rs vector threshold
  comparison — compare in f64 space to avoid threshold mismatches
- Use epsilon (1e-12) instead of exact == 0.0 for cosine similarity
  near-zero norm check in landscape.rs
- Improve LLM compiler error diagnostics — show meaningful message
  when response body read fails
- Add 4 tests: multibyte summary truncation, near-zero cosine norm,
  adjacent temporal ranges, containment temporal ranges

+++ engine/crates/covalence-core/src/ingestion/pg_resolver.rs
@@ -216,7 +216,7 @@ impl PgResolver {
 
         if let Some(r) = row {
             let sim: f64 = r.get("sim");
-            if sim as f32 >= self.vector_threshold {
+            if sim >= self.vector_threshold as f64 {
                 let id: NodeId = r.get("id");
                 let canonical_name: String = r.get("canonical_name");
                 return Ok(Some(ResolvedEntity {
